-added logic to clear the pdf file from memory
