+++ app/src/main/java/ca/bc/gov/bchealth/ui/healthpass/HealthPassesFragment.kt
@@ -1,11 +1,9 @@
 package ca.bc.gov.bchealth.ui.healthpass
 
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.view.View
 import android.widget.TextView
-import androidx.core.content.FileProvider
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
@@ -18,6 +16,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import ca.bc.gov.bchealth.R
 import ca.bc.gov.bchealth.databinding.FragmentHelathPassesBinding
+import ca.bc.gov.bchealth.utils.PdfHelper
 import ca.bc.gov.bchealth.utils.viewBindings
 import ca.bc.gov.bchealth.viewmodel.PdfDecoderViewModel
 import ca.bc.gov.bchealth.viewmodel.SharedViewModel
@@ -39,6 +38,12 @@ class HealthPassesFragment : Fragment(R.layout.fragment_helath_passes) {
     private val sharedViewModel: SharedViewModel by activityViewModels()
     private var patientId: Long = -1L
     private val pdfDecoderViewModel: PdfDecoderViewModel by viewModels()
+    private var fileInMemory: File? = null
+    private var resultListener = registerForActivityResult(
+        ActivityResultContracts.StartActivityForResult()
+    ) { _ ->
+        fileInMemory?.delete()
+    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -116,7 +121,8 @@ class HealthPassesFragment : Fragment(R.layout.fragment_helath_passes) {
                         val (federalTravelPass, file) = uiState.pdf
                         if (file != null) {
                             try {
-                                showPDF(file)
+                                fileInMemory = file
+                                PdfHelper().showPDF(file, requireActivity(), resultListener)
                             } catch (e: Exception) {
                                 navigateToViewTravelPass(federalTravelPass)
                             }
@@ -128,18 +134,6 @@ class HealthPassesFragment : Fragment(R.layout.fragment_helath_passes) {
                 }
             }
         }
-    }
-
-    private fun showPDF(file: File) {
-        val authority =
-            requireActivity().applicationContext.packageName.toString() +
-                ".fileprovider"
-        val uriToFile: Uri =
-            FileProvider.getUriForFile(requireActivity(), authority, file)
-        val shareIntent = Intent(Intent.ACTION_VIEW)
-        shareIntent.setDataAndType(uriToFile, "application/pdf")
-        shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-        requireActivity().startActivity(shareIntent)
     }
 
     private fun navigateToViewTravelPass(federalTravelPass: String) {
